separe clock constant
